+++ tool/generate-native-config.rb
@@ -48,8 +48,6 @@ when /aarch64-linux/
 when /x86_64-linux/
   PLATFORM_FILE = 'org/truffleruby/platform/LinuxAMD64NativeConfiguration.java'
   EXTRA_CFLAGS << ' -D_GNU_SOURCE'
-when /x86_64-darwin/
-  PLATFORM_FILE = 'org/truffleruby/platform/DarwinAMD64NativeConfiguration.java'
 when /arm64.*-darwin/
   PLATFORM_FILE = 'org/truffleruby/platform/DarwinAArch64NativeConfiguration.java'
 else